refactor: add a folder for interfaces
