+++ src/interfaces/CharacterInterface.ts
